requirement.txt
absl-py==1.0.0
alembic==1.8.1
amqp==5.1.1
anyio==3.6.1
appdirs==1.4.4
apturl==0.5.2
argh==0.26.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==22.1.0
autopep8==1.6.0
backcall==0.2.0
backports.weakref==1.0.post1
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.10.0
billiard==3.6.4.0
bkcharts==0.2
black==19.10b0
bleach==5.0.1
blinker==1.4
boto==2.49.0
Bottleneck==1.3.2
Brlapi==0.7.0
brotlipy==0.7.0
build==0.9.0
cached-property==1.5.1
cachelib==0.9.0
cachetools==4.2.4
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
command-not-found==0.3
contextlib2==0.6.0.post1
contourpy==1.0.5
cryptography==38.0.1
cupshelpers==1.0
cycler==0.10.0
cytoolz==0.11.0
dbus-python==1.2.16
debugpy==1.6.2
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.6
django-ckeditor==6.3.0
django-js-asset==2.0.0
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
dpcpp-cpp-rt==2022.2.0
duplicity==0.8.12.0
elasticsearch==7.15.1
entrypoints==0.3
et-xmlfile==1.0.1
executing==0.9.1
f==0.0.1
fastcache==1.1.0
fasteners==0.14.1
fastjsonschema==2.16.2
filelock==3.8.0
Flask==2.2.2
Flask-Caching==2.0.1
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
flask-redis==0.4.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==4.11.1
flatbuffers==2.0
fonttools==4.37.4
fpdf==1.7.2
fpdf2==2.5.0
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.3
grpcio==1.43.0
gunicorn==20.1.0
h11==0.14.0
h5py==2.10.0
HeapDict==1.0.1
helpdev==0.7.1
html5lib==1.1
httplib2==0.14.0
idna==3.4
imageio==2.22.2
importlib-metadata==4.9.0
importlib-resources==5.9.0
iniconfig==1.1.1
insta-scrape==2.1.2
instagramy==4.4
instascrape==0.1
intel-cmplr-lib-rt==2022.2.0
intel-cmplr-lic-rt==2022.2.0
intel-opencl-rt==2022.2.0
intel-openmp==2022.2.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isort==5.10.1
itsdangerous==2.1.2
jdcal==1.4.1
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.5
jsonschema==4.16.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.21.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.7.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.4
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==12.0.0
llvmlite==0.36.0
locket==0.2.1
lockfile==0.12.2
louis==3.12.0
lxml==4.7.1
ma==1.0.2
ma.mongo==1.0.2
macaroonbakery==1.3.1
Mako==1.2.3
Markdown==3.3.6
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
matplotlib==3.6.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==2.0.4
mkl==2022.2.0
mkl-fft==1.3.0
monotonic==1.5
more-itertools==4.2.0
mpmath==1.2.1
multipledispatch==0.6.0
mypy-extensions==0.4.3
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.28
mysqlclient==2.1.0
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.5
netifaces==0.10.4
networkx==2.8.7
notebook==6.5.1
notebook_shim==0.2.0
numexpr==2.8.3
numpy==1.22.1
oauthlib==3.1.1
olefile==0.46
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.2.4
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pathspec==0.7.0
patsy==0.5.1
pbr==5.10.0
pep517==0.13.0
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pip-tools==6.10.0
pipenv==11.9.0
pkginfo==1.7.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
pluggy==0.13.1
ply==3.11
prometheus-client==0.15.0
prompt-toolkit==3.0.30
protobuf==3.6.1
psutil==5.9.1
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.8.0
pycosat==0.6.3
pycparser==2.21
pycups==1.9.73
Pygments==2.12.0
PyGObject==3.36.0
PyJWT==2.5.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pytest==6.2.3
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.21.0
pytz==2022.1
PyWavelets==1.4.1
pyxdg==0.26
PyYAML==5.4.1
pyzmq==23.2.0
QDarkStyle==2.8.1
qtconsole==5.3.2
QtPy==1.9.0
redis==4.3.4
regex==2022.9.13
reportlab==3.5.34
requests==2.27.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.8
scikit-image==0.18.1
scipy==1.9.2
SecretStorage==2.3.1
selenium==4.4.3
Send2Trash==1.8.0
simplegeneric==0.8.1
simplejson==3.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.41
sqlparse==0.4.2
stack-data==0.3.0
stevedore==4.0.0
systemd-python==234
tables==3.6.1
tbb==2021.7.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.9.4
texttable==1.6.2
tifffile==2020.10.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
tweepy==4.5.0
typed-ast==1.5.4
typing_extensions==4.4.0
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unicodecsv==0.14.1
urllib3==1.26.12
usb-creator==0.3.7
validate-email==1.3
vine==5.0.0
virtualenv==20.16.6
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.4
wadllib==1.3.3
wcwidth==0.2.5
webdriver-manager==3.8.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
widgetsnbextension==3.5.1
wrapt==1.14.1
wsproto==1.2.0
WTForms==3.0.1
xkit==0.0.0
xlwt==1.3.0
xmltodict==0.12.0
zict==2.0.0
zipp==3.8.1
zope.event==4.5.0
